rename: MemberLocation 관련 dto 패키지 분리
